Changed IntComputer to accept large numbers (long)

+++ src/main/java/day5/SolverDay5.java
@@ -6,7 +6,7 @@ import java.util.List;
 
 public class SolverDay5 {
 
-    public int solve(List<Integer> numbers, int inputCode) {
+    public long solve(List<Long> numbers, int inputCode) {
         IntComputer intComputer = new IntComputer.IntComputerBuilder(numbers)
                 .inputCode(inputCode)
                 .phaseSetting(inputCode)
@@ -14,7 +14,7 @@ public class SolverDay5 {
         intComputer.start();
         intComputer.setRunning();
         intComputer.waitForIntcomputer();
-        List<Integer> intComputerOutput = intComputer.getIntComputerOutput();
+        List<Long> intComputerOutput = intComputer.getIntComputerOutput();
         return intComputerOutput.get(intComputerOutput.size() - 1);
     }
